[LEMS-2851/answerless-numeric-input] split out deriveAnswerForms
