+++ packages/perseus-core/src/data-schema.ts
@@ -1205,6 +1205,7 @@ export type PerseusNumericInputWidgetOptions = {
     // Used by examples, maybe not used and should be removed in the future
     // see TODO in numeric-input
     answerForms?: ReadonlyArray<PerseusNumericInputAnswerForm>;
+    fullAnswerForms?: ReadonlyArray<PerseusNumericInputAnswerForm>;
 };
 
 export type PerseusNumericInputAnswer = {
