+++ packages/perseus/src/widgets/numeric-input/numeric-input.test.ts
@@ -508,201 +508,4 @@ describe("transform", () => {
             ],
         });
     });
-
-    it("ignores the legacy answerForms", () => {
-        const transform = NumericInputWidgetExport.transform;
-        const widgetOptions: PerseusNumericInputWidgetOptions = {
-            coefficient: false,
-            static: false,
-            size: "normal",
-            answers: [],
-            answerForms: [
-                {
-                    simplify: "required",
-                    name: "pi",
-                },
-            ],
-        };
-        const renderProps = transform(widgetOptions);
-        expect(renderProps).toEqual({
-            coefficient: false,
-            static: false,
-            size: "normal",
-            answerForms: [],
-        });
-    });
-
-    it("deduplicates answer forms", () => {
-        const transform = NumericInputWidgetExport.transform;
-        const widgetOptions: PerseusNumericInputWidgetOptions = {
-            coefficient: false,
-            static: false,
-            size: "normal",
-            answers: [
-                {
-                    status: "correct",
-                    maxError: null,
-                    strict: true,
-                    value: 0.5,
-                    simplify: "required",
-                    answerForms: ["proper"],
-                    message: "",
-                },
-                {
-                    status: "correct",
-                    maxError: null,
-                    strict: true,
-                    value: 7.5,
-                    simplify: "required",
-                    answerForms: ["proper"],
-                    message: "",
-                },
-            ],
-        };
-        const renderProps = transform(widgetOptions);
-        expect(renderProps).toEqual({
-            coefficient: false,
-            static: false,
-            size: "normal",
-            answerForms: [
-                {
-                    simplify: "required",
-                    name: "proper",
-                },
-            ],
-        });
-    });
-
-    it("handles multiple answer forms", () => {
-        const transform = NumericInputWidgetExport.transform;
-        const widgetOptions: PerseusNumericInputWidgetOptions = {
-            coefficient: false,
-            static: false,
-            size: "normal",
-            answers: [
-                {
-                    status: "correct",
-                    maxError: null,
-                    strict: true,
-                    value: 0.5,
-                    simplify: "required",
-                    answerForms: ["proper"],
-                    message: "",
-                },
-                {
-                    status: "correct",
-                    maxError: null,
-                    strict: true,
-                    value: 7.5,
-                    simplify: "optional",
-                    answerForms: ["pi"],
-                    message: "",
-                },
-            ],
-        };
-        const renderProps = transform(widgetOptions);
-        expect(renderProps).toEqual({
-            coefficient: false,
-            static: false,
-            size: "normal",
-            answerForms: [
-                {
-                    simplify: "required",
-                    name: "proper",
-                },
-                {
-                    simplify: "optional",
-                    name: "pi",
-                },
-            ],
-        });
-    });
-
-    it("handles multiple simplify options with same math formats", () => {
-        const transform = NumericInputWidgetExport.transform;
-        const widgetOptions: PerseusNumericInputWidgetOptions = {
-            coefficient: false,
-            static: false,
-            size: "normal",
-            answers: [
-                {
-                    status: "correct",
-                    maxError: null,
-                    strict: true,
-                    value: 0.5,
-                    simplify: "required",
-                    answerForms: ["pi"],
-                    message: "",
-                },
-                {
-                    status: "correct",
-                    maxError: null,
-                    strict: true,
-                    value: 7.5,
-                    simplify: "optional",
-                    answerForms: ["pi"],
-                    message: "",
-                },
-            ],
-        };
-        const renderProps = transform(widgetOptions);
-        expect(renderProps).toEqual({
-            coefficient: false,
-            static: false,
-            size: "normal",
-            answerForms: [
-                // uses the first simplify option when two answers
-                // have the same form
-                {
-                    simplify: "required",
-                    name: "pi",
-                },
-            ],
-        });
-    });
-
-    it("handles multiple math formats with same simplify options", () => {
-        const transform = NumericInputWidgetExport.transform;
-        const widgetOptions: PerseusNumericInputWidgetOptions = {
-            coefficient: false,
-            static: false,
-            size: "normal",
-            answers: [
-                {
-                    status: "correct",
-                    maxError: null,
-                    strict: true,
-                    value: 7.5,
-                    simplify: "required",
-                    answerForms: ["mixed"],
-                    message: "",
-                },
-                {
-                    status: "correct",
-                    maxError: null,
-                    strict: true,
-                    value: 0.5,
-                    simplify: "required",
-                    answerForms: ["pi"],
-                    message: "",
-                },
-            ],
-        };
-        const renderProps = transform(widgetOptions);
-        expect(renderProps).toEqual({
-            coefficient: false,
-            static: false,
-            size: "normal",
-            answerForms: [
-                {
-                    simplify: "required",
-                    name: "mixed",
-                },
-                {
-                    simplify: "required",
-                    name: "pi",
-                },
-            ],
-        });
-    });
 });
