+++ packages/perseus/src/widgets/numeric-input/utils.test.ts
@@ -1,8 +1,16 @@
 import {mockStrings} from "../../strings";
 
-import {generateExamples, shouldShowExamples, unionAnswerForms} from "./utils";
+import {
+    deriveAnswerForms,
+    generateExamples,
+    shouldShowExamples,
+    unionAnswerForms,
+} from "./utils";
 
-import type {PerseusNumericInputAnswerForm} from "@khanacademy/perseus-core";
+import type {
+    PerseusNumericInputAnswerForm,
+    PerseusNumericInputWidgetOptions,
+} from "@khanacademy/perseus-core";
 
 describe("generateExamples", () => {
     it("returns an array of examples", () => {
@@ -189,5 +197,266 @@ describe("unionAnswerForms", () => {
 
         // Assert
         expect(result).toEqual(expected);
+    });
+});
+
+describe("deriveAnswerForms", () => {
+    it("passes through fullAnswerForms", () => {
+        // Arrange
+        const widgetOptions: PerseusNumericInputWidgetOptions = {
+            coefficient: false,
+            static: false,
+            size: "normal",
+            answers: [],
+            fullAnswerForms: [
+                {
+                    simplify: "required",
+                    name: "proper",
+                },
+            ],
+        };
+
+        // Act
+        const result = deriveAnswerForms(widgetOptions);
+
+        // Assert
+        expect(result).toEqual([
+            {
+                simplify: "required",
+                name: "proper",
+            },
+        ]);
+    });
+
+    it("only uses answer forms from correct answers", () => {
+        // Arrange
+        const widgetOptions: PerseusNumericInputWidgetOptions = {
+            coefficient: false,
+            static: false,
+            size: "normal",
+            answers: [
+                {
+                    status: "correct",
+                    maxError: null,
+                    strict: true,
+                    value: 0.5,
+                    simplify: "required",
+                    answerForms: ["proper"],
+                    message: "",
+                },
+                {
+                    status: "wrong",
+                    maxError: null,
+                    strict: true,
+                    value: 0.5,
+                    simplify: "required",
+                    answerForms: ["decimal"],
+                    message: "",
+                },
+            ],
+        };
+
+        // Act
+        const result = deriveAnswerForms(widgetOptions);
+
+        // Assert
+        expect(result).toEqual([
+            {
+                simplify: "required",
+                name: "proper",
+            },
+        ]);
+    });
+
+    it("ignores the legacy answerForms", () => {
+        // Arrange
+        const widgetOptions: PerseusNumericInputWidgetOptions = {
+            coefficient: false,
+            static: false,
+            size: "normal",
+            answers: [],
+            answerForms: [
+                {
+                    simplify: "required",
+                    name: "pi",
+                },
+            ],
+        };
+
+        // Act
+        const result = deriveAnswerForms(widgetOptions);
+
+        // Assert
+        expect(result).toEqual([]);
+    });
+
+    it("deduplicates answer forms", () => {
+        // Arrange
+        const widgetOptions: PerseusNumericInputWidgetOptions = {
+            coefficient: false,
+            static: false,
+            size: "normal",
+            answers: [
+                {
+                    status: "correct",
+                    maxError: null,
+                    strict: true,
+                    value: 0.5,
+                    simplify: "required",
+                    answerForms: ["proper"],
+                    message: "",
+                },
+                {
+                    status: "correct",
+                    maxError: null,
+                    strict: true,
+                    value: 7.5,
+                    simplify: "required",
+                    answerForms: ["proper"],
+                    message: "",
+                },
+            ],
+        };
+
+        // Act
+        const result = deriveAnswerForms(widgetOptions);
+
+        // Assert
+        expect(result).toEqual([
+            {
+                simplify: "required",
+                name: "proper",
+            },
+        ]);
+    });
+
+    it("handles multiple answer forms", () => {
+        // Arrange
+        const widgetOptions: PerseusNumericInputWidgetOptions = {
+            coefficient: false,
+            static: false,
+            size: "normal",
+            answers: [
+                {
+                    status: "correct",
+                    maxError: null,
+                    strict: true,
+                    value: 0.5,
+                    simplify: "required",
+                    answerForms: ["proper"],
+                    message: "",
+                },
+                {
+                    status: "correct",
+                    maxError: null,
+                    strict: true,
+                    value: 7.5,
+                    simplify: "optional",
+                    answerForms: ["pi"],
+                    message: "",
+                },
+            ],
+        };
+
+        // Act
+        const result = deriveAnswerForms(widgetOptions);
+
+        // Assert
+        expect(result).toEqual([
+            {
+                simplify: "required",
+                name: "proper",
+            },
+            {
+                simplify: "optional",
+                name: "pi",
+            },
+        ]);
+    });
+
+    it("handles multiple simplify options with same math formats", () => {
+        // Arrange
+        const widgetOptions: PerseusNumericInputWidgetOptions = {
+            coefficient: false,
+            static: false,
+            size: "normal",
+            answers: [
+                {
+                    status: "correct",
+                    maxError: null,
+                    strict: true,
+                    value: 0.5,
+                    simplify: "required",
+                    answerForms: ["pi"],
+                    message: "",
+                },
+                {
+                    status: "correct",
+                    maxError: null,
+                    strict: true,
+                    value: 7.5,
+                    simplify: "optional",
+                    answerForms: ["pi"],
+                    message: "",
+                },
+            ],
+        };
+
+        // Act
+        const result = deriveAnswerForms(widgetOptions);
+
+        // Assert
+        expect(result).toEqual([
+            // uses the first simplify option when two answers
+            // have the same form
+            {
+                simplify: "required",
+                name: "pi",
+            },
+        ]);
+    });
+
+    it("handles multiple math formats with same simplify options", () => {
+        // Arrange
+        const widgetOptions: PerseusNumericInputWidgetOptions = {
+            coefficient: false,
+            static: false,
+            size: "normal",
+            answers: [
+                {
+                    status: "correct",
+                    maxError: null,
+                    strict: true,
+                    value: 7.5,
+                    simplify: "required",
+                    answerForms: ["mixed"],
+                    message: "",
+                },
+                {
+                    status: "correct",
+                    maxError: null,
+                    strict: true,
+                    value: 0.5,
+                    simplify: "required",
+                    answerForms: ["pi"],
+                    message: "",
+                },
+            ],
+        };
+
+        // Act
+        const result = deriveAnswerForms(widgetOptions);
+
+        // Assert
+        expect(result).toEqual([
+            {
+                simplify: "required",
+                name: "mixed",
+            },
+            {
+                simplify: "required",
+                name: "pi",
+            },
+        ]);
     });
 });
