+++ packages/perseus/src/widgets/numeric-input/utils.ts
@@ -1,5 +1,8 @@
 import type {PerseusStrings} from "../../strings";
-import type {PerseusNumericInputAnswerForm} from "@khanacademy/perseus-core";
+import type {
+    PerseusNumericInputAnswerForm,
+    PerseusNumericInputWidgetOptions,
+} from "@khanacademy/perseus-core";
 
 /**
  * The full list of available strings for the numeric input widget,
@@ -118,4 +121,27 @@ export const unionAnswerForms: (
             formExampleKeys.indexOf(a.name) - formExampleKeys.indexOf(b.name)
         );
     });
+};
+
+export const deriveAnswerForms = (
+    options: PerseusNumericInputWidgetOptions,
+) => {
+    if (options.fullAnswerForms) {
+        return options.fullAnswerForms;
+    }
+
+    return unionAnswerForms(
+        // Filter out the correct answers and map them to the answer forms
+        // so that we can generate the examples for the widget.
+        (options.answers || [])
+            .filter((answer) => answer.status === "correct")
+            .map((answer) => {
+                return (answer.answerForms || []).map((form) => {
+                    return {
+                        simplify: answer.simplify,
+                        name: form,
+                    };
+                });
+            }),
+    );
 };
